+++ Veritas/src/main/scala/de/tu_darmstadt/veritas/VerificationInfrastructure/lemmagen/LemmaGenSpecEnquirer.scala
@@ -23,16 +23,23 @@ class LemmaGenSpecEnquirer(spec: Module, dsk: DomainSpecificKnowledge) extends V
       .toSeq
   }
 
-  /** Retrieve all boolean functions that take typ */
-  def retrievePredicates(typ: SortRef): Set[FunctionDef] = predicates.filter(_.signature.in.contains(typ))
+  /** Retrieve all boolean functions that take any of `types` */
+  def retrievePredicates(types: Set[SortRef]): Set[FunctionDef] =
+    types.flatMap(typ => predicates.filter(_.signature.in.contains(typ)))
 
-  /** Return all static and dynamic functions that take typ */
-  def retrieveTransformers(typ: SortRef): Set[FunctionDef] = functions.filter(_.signature.in.contains(typ))
+  /** Return all static and dynamic functions that take any of `types` */
+  def retrieveTransformers(types: Set[SortRef]): Set[FunctionDef] =
+    types.flatMap(typ => functions.filter(_.signature.in.contains(typ)))
 
-  /** Return all static and dynamic functions that produce a type involving typ, or typ itself */
-  def retrieveProducers(typ: SortRef): Set[FunctionDef] = {
-    val involvingTyp = typ +: getDataTypesInvolving(typ)
-    functions.filter(involvingTyp contains _.signature.out)
+  /**
+    * Return all static and dynamic functions that produce any type involving
+    * any of `types`, or any member of `types` itself
+    * */
+  def retrieveProducers(types: Set[SortRef]): Set[FunctionDef] = {
+    types.flatMap(typ => {
+      val involvingTyp = typ +: getDataTypesInvolving(typ)
+      functions.filter(involvingTyp contains _.signature.out)
+    })
   }
 
   def isFailableType(typ: SortRef): Boolean = dsk.failableTypes.exists(_.name == typ.name)
